types: extract app bootstrap types to app.ts
